extract workflow history routes
